requirements.txt
python-telegram-bot==20.7
gspread==6.2.0
oauth2client==4.1.3
python-dotenv
google-generativeai
